Defined ITagDomain and added Domain property to ITagType Implemented TagDomain with String, StringArray, PositiveInteger, Date and Millisecond types Changed SQLiteTagRepository and FirebirdTagRepository to use ITagDomain GetValue() method Implemented TagFactory

+++ trunk/Gnosis/src/Gnosis.Core/ITagSchema.cs
@@ -7,7 +7,7 @@ namespace Gnosis.Core
 {
     public interface ITagSchema
     {
-        long Id { get; }
+        int Id { get; }
         string Name { get; }
     }
 }
